Add support for extended position vertex buffer.

+++ src/shaders/RSPProcessCS.hlsl
@@ -19,8 +19,8 @@ struct RSPProcessCB {
 };
 
 [[vk::push_constant]] ConstantBuffer<RSPProcessCB> gConstants : register(b0);
-Buffer<int> srcPos : register(t1);
-Buffer<int> srcVel : register(t2);
+Buffer<float> srcPos : register(t1);
+Buffer<float> srcVel : register(t2);
 Buffer<float> srcTc : register(t3);
 Buffer<float> srcTcVel : register(t4);
 Buffer<uint> srcCol : register(t5);
